Butterfly/butterflyApp/consumers/telegram-consumer/requirements.txt
kafka-python==1.4.6
mongoengine==0.17.0
pyTelegramBotAPI==3.6.6